feat: add bookmark menu item, rename starred pages, fix review issues
- Add "Bookmark Page" action to reader menu Navigation section
- Rename "Starred Pages" to "Bookmarks" across all 18 languages
- Fix FileBrowser page navigation (onNextContinuous/onPreviousContinuous)
- Fix RecentBooks bold title width measurement
- Fix HalGPIO uint16_t truncation in verifyPowerButtonWakeup
- Align Lexend line spacing with Bookerly/Bokerlam
- Add clockMode to settings list for persistence

+++ lib/hal/HalGPIO.cpp
@@ -264,8 +264,8 @@ void HalGPIO::verifyPowerButtonWakeup(uint16_t requiredDurationMs, bool shortPre
     return;
   }
   // Calibrate: subtract boot time already elapsed, assuming button held since boot
-  const uint16_t calibration = millis();
-  const uint16_t calibratedDuration = (calibration < requiredDurationMs) ? (requiredDurationMs - calibration) : 1;
+  const unsigned long calibration = millis();
+  const unsigned long calibratedDuration = (calibration < requiredDurationMs) ? (requiredDurationMs - calibration) : 1;
 
   const auto start = millis();
   inputMgr.update();
